bind/capps-named: Test if BIND loads zones, responsds properly to query
The test starts up capps-named using dummy zone files and runs a DNS
query against the loaded zones. The test is deemed successful if response
contains the expected SOA record.

Add bind-tools to Docker image. Test uses dig to query test DNS zones.

Running capps-named no longer returns BIND version, as it was necessary
to remove the -v arg for the other parts of the test.

The config directory no longer mounts read-only, as BIND will not run
unless the working directory is writable.

BIND requires that the user running named either owns the working
directory or belongs to the group owning it. Because Docker carries over
file ownership from host system to container, it is necessary to do
Docker run with a GID derived from /tconfig. The user is set to root
because named requires root privileges to access port 53.