bind/capps-named: Test if BIND loads zones, responsds properly to query
The test starts up capps-named using dummy zone files and runs a DNS
query against the loaded zones. The test is deemed successful if response
contains a string from the test config.

Running capps-named no longer returns BIND version, as it was necessary
to remove the -v arg for the other parts of the test.

The config directory no longer mounts read-only, as BIND will not run
unless the working directory is writable.

BIND requires that the user running named (1) be the root user (in order
to access privileged ports), and (2) have write access to the working
directory. For this reason, the Docker container is started as user root
with the GID of the local directory being bound to the container.