bind/capps-named: Add bind-tools to Docker image, add DNS query to named server in test
Test DNS query uses zone files in test-config. Query is deemed successful if
response includes the appropriate SOA record.

Explicitly shut down Docker in Test, as it was necessary to remove -v arg when
evoking capps-named. This also means that BIND version cannot be verified in
the same way as before. Need to re-implement this or find a way around the
absent -v arg.

bind-tools was added to Docker image because test query uses dig.