bind/capps-named: Run container as root:gid-of-tconfig
The ownership of /tconfig will be inherited by the container volume
regardless of whether or not a corresponding group exists in the
container. By getting gid of /tconfig and running container as that
group, we are satisfying BIND's permission requirement.

BIND does not seem to mind which user runs named, so long as said user
has write access and belongs to the group owning the working directory.
Root is required to utilise port 53, so is an approprate user to start
the container with.

Had to mount volume as read-write, as named will not run if volume is
read-only.