bind/capps-named: Run in Python instead of bash
Replace bash script with Python script. The idea is to leverage Python's
abilities to manage Docker containers and DNS processes, and to bring
the DNS server in line with the test client (which is already Python).

Using docker-py to manage Docker processes, and dnspython to manage DNS
server and test client.

Testing no longer requires a .cid file to store the container ID; this
has been removed. The .build folder is now used exlusively by venv. Test
cleanup is handled by docker-py.

As of now, all options for capps-named are hard-coded; the next step is
to re-instate the options which were in the old script. For this reason,
no unit tests have been implemented yet. These too will come later.